cast to float before converting to numpy (#234)

+++ llm_studio/src/metrics/text_causal_language_modeling_metrics.py
@@ -146,7 +146,7 @@ class Perplexity(nn.Module):
 
 
 def perplexity(cfg: Any, results: Dict, val_df: pd.DataFrame):
-    return results["perplexity"].detach().cpu().numpy()
+    return results["perplexity"].detach().float().cpu().numpy()
 
 
 class Metrics:
